updated scripts

+++ bin/multimer/heteromer_post_af.py
@@ -119,8 +119,9 @@ def main(argv):
                 os.system(f"sed -i 's/>{processed_seuqences[monomer_sequence]}/>{monomer_id}/g' " + os.path.join(N7_monomer_outdir, 'msa', msa))
             monomer_qas_res[monomer_id] = copy.deepcopy(monomer_qas_res[processed_seuqences[monomer_sequence]])
 
-
-    bash_script_dir = os.path.join(N6_outdir, 'post_def_bash_scripts')
+    bash_script_dir = os.path.join(N3_outdir, 'post_def_bash_scripts')
+    if os.path.exists(params['slurm_script_template']):
+        bash_script_dir = os.path.join(N3_outdir, 'post_def_slurm_scripts')
     os.makedirs(bash_script_dir, exist_ok=True)
 
     run_methods = ['folds_iter', 'folds_iter_nop', 'folds_iter_not', 'folds_iter_esm', 'folds_iter_esm_nop',
@@ -130,10 +131,22 @@ def main(argv):
         cmd = f"python bin/multimer/heteromer_foldseek.py --option_file {FLAGS.option_file} " \
               f"--fasta_path {FLAGS.fasta_path} --output_dir {FLAGS.output_dir} " \
               f"--config_name {run_method}"
-        bash_file = os.path.join(bash_script_dir, run_method + '.sh')
-        print(f"Generating bash files for {run_method}: {bash_file}")
-        with open(bash_file, 'w') as fw:
-            fw.write(cmd)
+
+        if os.path.exists(params['slurm_script_template']):
+            bash_file = os.path.join(bash_script_dir, run_method + '.sh')
+            print(f"Generating bash file for {run_method}: {bash_file}")
+            jobname = f"{targetname}_{run_method}"
+            with open(bash_file, 'w') as fw:
+                for line in open(params['slurm_script_template']):
+                    line = line.replace("JOBNAME", jobname)
+                    fw.write(line)
+                fw.write(cmd)
+            # os.system(f"sbatch {bash_file}")
+        else:
+            bash_file = os.path.join(bash_script_dir, run_method + '.sh')
+            print(bash_file)
+            with open(bash_file, 'w') as fw:
+                fw.write(cmd)
